% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_con.R
\name{connect_test}
\alias{connect_test}
\title{Test Database Connection}
\usage{
connect_test(driver, port, dbname, host)
}
\arguments{
\item{driver}{A DBI-compatible database driver (e.g. \code{RPostgres::Postgres()}).}

\item{port}{Integer. Database port.}

\item{dbname}{Character. Name of the database.}

\item{host}{Character. Database host address.}
}
\value{
Logical. \code{TRUE} if the connection succeeds, otherwise \code{FALSE}.
}
\description{
Attempts to establish a temporary database connection using credentials
stored in environment variables and immediately disconnects if successful.
}
\details{
This function uses the environment variables \code{DBTools_db_user} and
\code{DBTools_db_password} for authentication.
}
\examples{
\dontrun{
connect_test(
  driver = RPostgres::Postgres(),
  port = 5432,
  dbname = "forschung",
  host = "10.82.152.222"
)
}

}
