% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTables.R
\name{listTables}
\alias{listTables}
\title{Download  data from PostgreSQL-Database}
\usage{
listTables(dbSchema, db = "forschung", connection = NULL, ...)
}
\arguments{
\item{dbSchema}{character object, name of the database scheme}

\item{db}{character object, name of the database}

\item{connection}{pool connection, optional, a pool connection e.g. created with db_con(as_pool = TRUE)}

\item{...}{arguments passed to db_con()}
}
\value{
A dataframe with 2 columns - schema name and table names
}
\description{
This function creates a connection to the Nationalpark Hunsrück-Hochwald
PostgreSQL-database and downloads a specific table with or without a geometry
column by using postgis
}
